fix: auto-resolve company logos in list view, not just modal
Move KNOWN_COMPANIES map and lookupCompanyDomain() into CompanyLogo
component so it auto-resolves domain from company name everywhere.
Previously only the modal did the lookup — list items with no saved
companyDomain showed the fallback initial even for known companies.
