+++ app/form/onboarding/components/experience/AddWorkExperienceModal.tsx
@@ -14,106 +14,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { CompanyLogo } from "@/components/ui/company-logo";
+import { CompanyLogo, lookupCompanyDomain } from "@/components/ui/company-logo";
 import { WorkExperience } from "./WorkExperienceSection";
-
-// Well-known company name → domain mapping for auto-detection
-const KNOWN_COMPANIES: Record<string, string> = {
-  google: "google.com",
-  microsoft: "microsoft.com",
-  amazon: "amazon.com",
-  meta: "meta.com",
-  facebook: "meta.com",
-  apple: "apple.com",
-  netflix: "netflix.com",
-  salesforce: "salesforce.com",
-  oracle: "oracle.com",
-  ibm: "ibm.com",
-  accenture: "accenture.com",
-  deloitte: "deloitte.com",
-  mckinsey: "mckinsey.com",
-  "goldman sachs": "goldmansachs.com",
-  "jp morgan": "jpmorgan.com",
-  jpmorgan: "jpmorgan.com",
-  stripe: "stripe.com",
-  uber: "uber.com",
-  airbnb: "airbnb.com",
-  spotify: "spotify.com",
-  adobe: "adobe.com",
-  cisco: "cisco.com",
-  vmware: "vmware.com",
-  snowflake: "snowflake.com",
-  databricks: "databricks.com",
-  palantir: "palantir.com",
-  coinbase: "coinbase.com",
-  tesla: "tesla.com",
-  nvidia: "nvidia.com",
-  intel: "intel.com",
-  samsung: "samsung.com",
-  twitter: "x.com",
-  x: "x.com",
-  linkedin: "linkedin.com",
-  github: "github.com",
-  atlassian: "atlassian.com",
-  slack: "slack.com",
-  zoom: "zoom.us",
-  shopify: "shopify.com",
-  twilio: "twilio.com",
-  square: "squareup.com",
-  block: "block.xyz",
-  paypal: "paypal.com",
-  razorpay: "razorpay.com",
-  flipkart: "flipkart.com",
-  swiggy: "swiggy.com",
-  zomato: "zomato.com",
-  infosys: "infosys.com",
-  tcs: "tcs.com",
-  "tata consultancy services": "tcs.com",
-  wipro: "wipro.com",
-  hcl: "hcltech.com",
-  reliance: "ril.com",
-  jio: "jio.com",
-  paytm: "paytm.com",
-  phonepe: "phonepe.com",
-  byju: "byjus.com",
-  "byju's": "byjus.com",
-  ola: "olacabs.com",
-  meesho: "meesho.io",
-  cred: "cred.club",
-  dream11: "dream11.com",
-  zerodha: "zerodha.com",
-  groww: "groww.in",
-  unacademy: "unacademy.com",
-  notion: "notion.so",
-  figma: "figma.com",
-  vercel: "vercel.com",
-  netlify: "netlify.com",
-  supabase: "supabase.com",
-  mongodb: "mongodb.com",
-  elastic: "elastic.co",
-  datadog: "datadoghq.com",
-  cloudflare: "cloudflare.com",
-  aws: "aws.amazon.com",
-  "amazon web services": "aws.amazon.com",
-  "google cloud": "cloud.google.com",
-  azure: "azure.microsoft.com",
-  "digital ocean": "digitalocean.com",
-  digitalocean: "digitalocean.com",
-  heroku: "heroku.com",
-  redhat: "redhat.com",
-  "red hat": "redhat.com",
-  docker: "docker.com",
-  hashicorp: "hashicorp.com",
-  jetbrains: "jetbrains.com",
-  postman: "postman.com",
-  gitlab: "gitlab.com",
-  bitbucket: "bitbucket.org",
-};
-
-function lookupCompanyDomain(companyName: string): string | null {
-  const normalized = companyName.trim().toLowerCase();
-  return KNOWN_COMPANIES[normalized] || null;
-}
 
 interface AddWorkExperienceModalProps {
   isOpen: boolean;
